+++ platform/src/sdk/models/operations/createapikey.ts
@@ -35,5 +35,5 @@ export class CreateApiKeyResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }