+++ platform/src/sdk/models/operations/createrule.ts
@@ -29,7 +29,7 @@ export class CreateRuleResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * OK