+++ platform/src/sdk/models/operations/getsupplementaldataconfiguration.ts
@@ -70,7 +70,7 @@ export class GetSupplementalDataConfigurationResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * OK