+++ platform/src/sdk/models/operations/listrules.ts
@@ -55,7 +55,7 @@ export class ListRulesResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * OK