+++ platform/src/sdk/models/operations/updateprofilesyncsettings.ts
@@ -57,5 +57,5 @@ export class UpdateProfileSyncSettingsResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }