+++ platform/RELEASES.md
@@ -69,3 +69,13 @@ Based on:
 - [typescript v2.0.0] platform
 ### Releases
 - [NPM v2.0.0] https://www.npmjs.com/package/@codat/platform/v/2.0.0 - platform
+
+## 2023-11-20 01:30:50
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.0.1] platform
+### Releases
+- [NPM v2.0.1] https://www.npmjs.com/package/@codat/platform/v/2.0.1 - platform